server: Move server Dockerfile to lib/server
Change-type: patch
