+++ docker-compose.dev.yml
@@ -1,7 +1,7 @@
 version: "3"
 services:
   web:
-    build: .
+    build: lib/server
     command: make start-server
     ports:
      - "8000:8000"
